Remove Equal method from the Signature type
